Skins2: change DTD to add possibility to store ID for Anchors, Panels and Groups.
It adds possibility to store IDs for Anchors, Panels and Groups.
This does not change the way skins2 works I believe. It just stops the
XML parser from complaining about non standard skins when they have set
IDs for those elements. And setting IDs for them is vital in the skin
editor, as it differentiates between different elements in the skin just
by their ID.