Remove dummy input locking
Contrary to commit 719f12a6 log, input_GetItem() does not check for a
lock on the input. There may be need for locking here(?), but taking the
object lock would definitely not help, as this was the only place is the
entire tree using the generic object lock on an input object.