Almost wait-free snapshot handling in the vout thread
Using vlc_mutex_trylock(), we make sure that the video output thread
will not yield because of an incomplete snapshot request. It can still
yield in the very unlikely case that a thread is doing a snapshot
request while the video output thread finishes sorting SPUs _and_ there
already was a pending snapshot earlier.