Try to use video-filter instead of vout-filter for deinterlacing.
It only works for blend and X mode but there are probably the most used.
When blend/X are enabled/disabled:
 - It avoids recreating the vout.
 - It avoids deinterlacing the (soft) subtitles.
The other modes are not possible without a lot of vout works, so it is
postponed for later.

As a side effect, the initial value of the menu now takes into account the
command line options.

Please, report any regression. This code is made complicated by the
vout filter/video filter2 switches&tricks.