codec/schroedinger: inform vlc of clean area
For SD, this should allow removal of black bars form the edge of frames.
i_aspect is set to the aspect ratio of the whole frame which may be
greater than that of the clean area.  VLC should then calculate the SAR,
and eventually display a clean area sized picture that is of the correct
aspect ratio.

NB, we do actually know the SAR, but it isn't known if vlc cares at this
point.

Signed-off-by: David Flynn <davidf@rd.bbc.co.uk>
Signed-off-by: Jean-Baptiste Kempf <jb@videolan.org>