libmpeg2: Fix Closed Captions in GOP.
This fixes a case where our libmpeg2 module was not extracting
Closed Captions from the user_data, when this user_data was in the
GOP header, instead of in the picture header. This information is
retained until the next picture (date) is handled.

See sample "bob-dvd-closedcaptioning works in ffmpeg but not in libmpeg2.ts"