Remove internal libmpeg2 from xvmc codec/vout
This removes deprecated argument from libmpeg2 functions.
Only mpeg2_internal.h remains used and needs to be updated,
but it can also be installed as a system header.

This also fixes warnings in xcommon.c.

Previously, this vout didn't work for me. (nVidia)
It is aimed to be tested on Intel/Via XvMC capable hardware.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>