WinCE: Require a recent enough mingwce to fix *printf problems
This just in: CeGCC has a MinGW with working replacement *printf
functions. Just like for mingw32, require (and allow building with) a
recent enough version of mingwce.

Since MinGW is patched by CeGCC, this check is not strictly sufficient;
but at least this allows to have working WinCE builds again.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>