Merge intf_Create() and intf_RunThread()
Interfaces do not have an event handler that would justify the
separation. Indeed, no caller did anything in-between.

Also, the "primary" reference to the interface object belongs to the
main thread because of the libvlc cleanup procedure. Therefore, the
interface thread pointer really should not be returned to the creator.
Note that this does not really fix the small race condition but it
conceals it within intf_Create() and libvlc_InternalCleanup().