avi demux: don't disable tracks when they end.
Disabling the tracks like this is an immediate action that messes up --file-caching for instance. There may be some cases where it is appropriate, and some cases where a track should be removed, but not when a track ends. In that case the core is responsible for cleanup.

Tested in bugfix for normal avi, broken avi and repaired avi, but not with "streamed" avi.

Should close #2853