Qt: FSC - remove the WIN32 special mode.
This makes Windows use the same mode with transparency than on Linux.
This removes the opacity trick to hide() the code.
This fix a few issues that were happening and difficult to fix because of this difficulty.

It will probably break stuffs for people on OLD windows version (2000,XP) with broken graphic cards... (Deactivate it by yourself) But I have spent enough time on this code.