skins2: solve crashes due to releasing variables at termination
skins was using two lists of variables (named and anonymous).
The problem was that variables from one list held references to variables from the other list and vice-versa. Whatever the order of releasing them, crashes could occur.

This patch uses the anonymous list to keep a reference on **all** variables.
This guarantees they are released in the reverse order from creation.