config: if read-only, refuse to save the configuration
This restores the pre-1.0.0 behaviour. With the atomic vlcrc update,
VLC managed to work-around a read-only vlcrc. Indeed, Linux (POSIX?)
allows renaming a file onto an existing read-only file, which is
then replaced (both content and meta-data). In principle, you should
put the containing directory in read-only mode to avoid this, but this
is inconvenient as .config/vlc contains other files.