mozilla: do not ship gui/macosx plugin in Safari webplugin
It is the main cause for memory leaks of VLC browser plugin for MacOS X. Every time the browser plugin is reloaded, vlc core reloads the plugins which causes all global defined objects to initialize. Since there are objects in libmacosx_plugin.dylib that are supposed to live as long as the application these are never freed. Testing is done with --enable-debug.