Fix video delay when using alsa or oss as slaves
Decrease poll delay from 0.5 to 0.01 seconds in alsa in oss, else the
audio input would block the video input when used as an input slave.
Other accesses likely to be used as input slaves might need to be
changed. An easier solution (module-wise, harder core wise) would be to
run each slave in its own thread.