Contribs: remove XML dependency from Fontconfig.
This removes around 1.1MB on both libass dll and freetype dll (on 2.8MB), which are 4th and 5th dll in size for Windows
(qt4, avcodec, avformat are bigger).
We DON'T use the fonts.conf anyway on Windows, therefore, why having such a dependency, just for parsing a configure file that we don't ship?
I still don't get what is wrong with an .ini/.conf format, but that is not my business, I guess.