freetype: use fontconfig to get default font if available
With fontconfig freetype-font is font-family to use, not fontfile.
This should solve issue on linux-side where there aren't any one
place and one file thats usually there.

I'm not that familiar with freetype/fontconfig stuff, so this one could
be way wrong. So feel free to mention/fix/revert this if you spot
something odd.