Added a new option, play-and-pause. When this option is set, the input thread is automatically paused when each item reaches the end of the file.
This option is intended to make it easier to use VLC during live
productions featuring instructors and speakers working
with American Sign Language (ASL), or really any Sign Language, where
it is common for a lecturer to refer to video clips during a
presentation. The automatic pausing provides a smooth pause for the
video operator to switch the projection system back to the camera
with the live speaker once a clip has finished playing.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>