vlc_detach: releases a thread handle asynchronously
Note that this can only be used safely in the core. In a plug-in, it
would introduce a race whereby dlclose() unmaps the code segment that
a detached thread is running (this was discussed over a year ago).
For that, we'd need something a bit more involved along the lines of
Win32's FreeLibraryAndExitThread().