freetype: force updating fontcache on loading of module
Remove external thread for fontdbbuiler, and do it on load. First run is
slower on windows, so show dialog about it to user. Following runs
should be much faster (atleast according to my own tests).

So on optimal case user sees that font-dialog only once when initial
cache is created.

One idea was to load freetype-module on start of vlc, so it would create
that cache on start of vlc instead of start of video (when showing
title), and could be changed to show something like 'VLC is starting
first time, please wait a while...' or so.

Feel free to revert/change/complain about this if it doesn't work for
you.