Qt4: store input_item_t* in plitem and handle metadata in model
No need to store/recreate metadata on plitem anymore, as we can get
it now on the fly, also separate headerData and Data. Hopefully
that vlc_gc_incref/vlc_gc_decref is sufficient for that input_item_t
pointer protection.

Feel free to fix/revert/comment if that pointer shouldn't really be
there, or if it's not safe it to be there thisway.