qt4: playlist: allow to hide all but the index 0 column
In some circumstances QTreeView does not behave properly if the column index 0 can hide/reappear.

Precisely, when column 0 is hidden and an item get's subitems, those subitems can not be D&D-ed.
If column 0 is then unhidden, trying to D&D the subitems will fail or result in unexpected behaviour.