XCB: allow building without xcb-utils
Of xcb-utils, we only use xcb-keysms at the moment. xcb-keysyms is only
needed for the globalhotkeys and xcb_window plugins. If there is another
window provider plugin, we can do without them so long as the user does
not set --no-embedded-video.

Note: xcb-utils is still needed for the LibVLC embedded video as it
depends on xcb_window. We could  #ifdef the key press support in
xcb_window if this is an issue (I didn't bother).