Qt4: remove (no-)sout-keep option
This is not really relevant as long as we have only one item.
However, sout-keep should really always be on anyway when streaming.
Most protocols would terminate client connections when the item
changes, and the file output would overwrite the file.

(sout_gather would be more problematic. We should use it too, but it
has always made sout_display crash.)