skins(Linux) : fix crash with a few skins (e.g Orangeade) and xcb_x11
On X11, windows are created with a 1x1 size and updated later on.
In some context, vout_display can catch this 1x1 size before the size is updated.
This results in a crash in swcaler.c which gets a 1x0 size (division by zero)
Changing default to 10x10 solves the problem.