opengllayer: Remove this module.
This hasn't been converted to "vout display".
This is to my knowledge not in use.
This is calling main thread synchronously on every redraw making the overhead too important, and the overall performance of the module very bad.
There is to my knowledge no way to fix this using public API on 10.5, without a complete rewrite.