skins2: improve deallocation of ressources for layouts and controls
Layouts and Controls are interrelated. Whatever the ones first deallocated, it leaves pointers referencing objects already destroyed. and potentially means memory leak.

This patch adds an unsetLayout() function to pair the setLayout() function and aimed at releasing resources.
Policy should now be that things allocated in constructor are released in destructor and things allocated in setLayout are released in unsetLayout.