bandlimited: avoid large stack allocation (refs #3199)
In most cases, there is enough space for the 2 old samples in the
input buffer head room. In other cases, we anyway need to memory copy
the whole buffer. So we now use block_Realloc(). This also saves us from
copying every samples when resampling.

Unfortunately, the transcode plugin seems to be feeding crap into the
resampler, thus it still crashes.