Load the command line into the LibVLC object variables (fixes #1941)...
...instead of the configuration. As a side effect, the command line
parameter should not be visible in the preferences, and more importantly
not be saved as part of the persistent configuration (fixes #1428)
(and really fixes older #1106).

We might be able to remove a few "dummy" change_unsaveable(). Some of
them really were just work-arounds for this bug. In principle, we could
possibly remove all of them as long as we keep add_internal().

Note that this commit will render any command line option inoperant if
it is read with config_Get*() instead of var_Inherit*() or var_*Get*().
I already fixed the most commonly used occurences, but there are some
left, especially inside plugins.