ASX ads behaviour
When the first entry is read and has a clientskip attribute, It keeps a
backup of the playlist entry for later use. If another entry exists
without clientskip attribute, the backup will be discarded.

When finished parsing, we search for a backup, if found, we add it (it
means we did not find any entry).