Simplify vcdx configure.ac logic.
HAVE_{LIBCDIO,LIBVCDINFO} and have_{libcdio,libvcdinfo} were nowhere
used in the source except when --enable-vcdx was given, hence only probe
for them when vcdx is to be enabled.
Since have_libvcdinfo (which is a requirement for vcdx) can only be
enabled if have_libcdio was defined, I simply put it in the enable_vcdx
block. And because vcdx needs to be explicitly enabled, we can safely
AC_MSG_ERROR on every missing library.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>