Allow vlc_object_attach to reparent directly
This saves one unlock/lock cycle, and makes sure that the reparented
object remains in the tree at all times. This does not fix the race in
reparenting, but it reduces its window of opportunity.

This also makes vlc_object_detach() essentially redumdant with
vlc_object_release() and vlc_object_attach().