Create preparser and fetcher immediately
They do not create threads anymore (unless they have work to do).

LibVLC needs this because we are sloppy programmers^W^W^W^Wit still
depends on the playlist for item meta infos (it should probably
instantiate the fetcher and preparser directly).

Also, do not create the preparser if there is no fetcher.