skins2: make default more consistent for video control
For a video control, default values are contradictory, in
a sense that a video control is not resizable and autoresize
is true. By default, this results in skins resizing to fit a new video, but
the inner video not using this extra space. (see AirFlow for instance)

This patch ensures that if autoresize is true, default behavior for a video
 control is to be resizeable.