Qt4: use explicit direct connections when passing VLC pointers...
...without reference. With a direct connection, the call stack ensures
that the pointer remains valid until the slot returns. This is not the
case for asynchronous ("queued" in Qt4 documentation) signals.

Obviously, this means we have to emit the signals from the Qt4 main loop
thread, *or* alternatively not touch any Qt4 UI stuff fromt he slot.
Hopefully, any such bug would be easier to find this ways.