luasd: Force a garbage collect at the end of Run().
Generally the core won't call Close() until the end of vlc's life. Hence
it's our last chance to gc collect before the end of vlc's life.

This gets important because some sd (freebox...) will open a vlc.stream
which (and that's an other issue) will eat 12MB at creation for a cache
that will never be completely filed. It also mean that if you vlc.stream
N small files, we may consume N * 12MB until finally lua decides to garabge
collect. And because we don't teach lua that vlc.stream takes 12MB this
is not likely to happen.