Work around cleanup handler bug in kfile plugin
Parental advisory: explicit hack.

If kfilemodule is loaded, bypass run cleanup handlers. Library really
should not use this C feature in any case (IMHO). And I find it
outright idiotic for a (Qt4) plugin to use cleanup handlers.
This is causing so many crash reports that it's not funny anymore:
https://bugs.launchpad.net/ubuntu/+source/vlc/+bug/408719