Remove the Xlib no-ARGB hack
The XCB/XVideo (for a long time) and the XCB/X11 (since the commit
before the previous one) video output plugins will fail safe in case
the embedding window has an alpha channel. In VLC 1.0, the Xlib plugins
needed this hack as the Xlib error handler would terminate VLC.

If compositing is in use, VLC will now use GLX. This will require
software chroma conversion from YUV to RGB. Also, unless the GLX output
is explicitly selected, VLC will flash twice before it starts rendering.
First, it tries (and fail) to use XVideo, then plain X11. To fix this
cosmetic problem, we would need to keep the window across plugins.

I assume this change is needed to fix #3581, but I have not tested, so
it may or may not be sufficient.