skins2: fix skins crashing when playlist is flat
m_firstPos (iterator associated to a stl list container) is not guaranteed to be valid (the associated item may have been deleted). Up to vlc1.0, this was no problem, because deleting an item was also buggy (a no-op). From vlc1.1 on, m_firstPos must now be recomputed to ensure it keeps on pointing to a valid item in all cases.

start for instance a playlist with the 'Dragon' skin as a test case.
This was pointed out by colt45 (from the forum)