Added an image demuxer.
It replaces the "fake" demuxer and decoder, and partially
the mjpeg one.
It does not rely on the extensions to detect the format.
The detected picture formats are: XCF, GIF, PNG, TIFF, BMP, PCX,
LBM, PNM, JFIF and SPIFF(Jpeg).
The reload feature of the fake:// demuxer is lost, as well as the deinterlace
feature of the fake decoder, patches are welcome.

It closes #3107 and #3388, and modifies #2295.