msw: remove unneeded code and useless "shaking" of window
For lack of a faster detection (based on win32 api messages),
resize/move of the parent window are detected in the Manage function
on a regular basis and acted upon if need be.

Move is an issue only for directx (surface doesn't automatically follow
the associated window). Though coordinates doesn't change, an UpdateRects
is still required to update the various rectangles used by directx.

As it was coded, everything was already taken care of.
The artificial "shaking" of the window was quite useless.