dbus control: Split the mpris dbus interfaces
The mpris defines three dbus interfaces with the same name.
This is wrong on many levels and the second version of the spec
which is currently being drafted fixes that.

In order to be able to reflect this change, the dbus control module
should not assume that these interface have the same name.

With this patch applied, VLC still complies with the mpris v1.0

Signed-off-by: Rafaël Carré <rafael.carre@gmail.com>