udev: use ALSA to find capture device names
This is a little bit more understandable than device numbers.
Unfortunately, hot plug is as flaky as before: udev reports the device
before it is ready. alsa-lib provides no event interface (that I know),
so we have to stick with udev for now though.

Also, alsa-lib is _not_ used to fetch the card name. In my opinion, the
udev vendor string is a lot better. Compare:
- "USB Device 0xccd:0x77" (alsa card name) with
  "TerraTec Electronic GmbH" (udev vendor), or
- "Intel ICH6" (alsa card name) with "Intel Corporation" (udev vendor).