LUA: remove unused catalog and DTD support
In my opinion, we do not want to encourage use of this.

On the one hand, our libxml2 catalog support has never been thread-safe
and this seems to have caused problems in the past (looking at the
skins2 comments). On the other hand, the xtag plugin never implemented
catalogs, so XML parsers have to handle unknown XML tags gracefully in
any case.

Also, shipping DTDs with VLC is awkward and arguably a waste of space.