Support DVD LPCM encode.
This patch adds encode support to the LPCM module; only the DVD variant for
now. It's not been tested with any other clients than VLC.

The rationale for this functionality is relatively simple -- LPCM is
currently the only low-latency codec that VLC supports and that can be
embedded in a TS stream. Longer-term support for other low-latency codecs
would be good, but if you're not particularly bandwidth-constrained and need
low latency, this is currently the best option I know of.

Signed-off-by: Rémi Denis-Courmont <remi@remlab.net>