lua: copy input options to result of playlist script
This fixes this kind of scenario: you want to save a video from your
favorite video website. So you open the VLC GUI, you copy/paste the URL
into the convert dialog and you click on okay. The corresponding
playlist lua script processes the URL and a new item is added into the
playlist, but the :sout option is lost in the process, and you start
watching a video that is never saved.