skins2(Linux): A XSync is needed
This partially fixes the position indicator getting blurry, because XGetImage didn't
retrieve the latest change in pixmap (Previous XCopyArea calls may still be pending)

Note that XSync is needed but not sufficient. Sporadic blurriness can still happen, because
there seems to be no way to know for sure when the pixmap content is really updated.
(waiting for NoExpose event, if enabled, is also not enough). Also, this problem is more
prone to occur with --vout xv than with x11 or opengl, as if xvideo were stressing the X server
more than the other two.

 modules/gui/skins2/x11/x11_graphics.cpp |    4 ++++
 1 files changed, 4 insertions(+), 0 deletions(-)