EGL: initial video output plugin
Currently, this require EGL 1.3 or later to compile.
This was only tested against the GLX driver from the Mesa Project,
and desktop OpenGL, which requires EGL 1.4. Win32 support is untested
(not even compilation).

OpenGL ES 1 and 2 are mostly implemented but completely untested too.
The USE_OPENGL_ES constant would have to be predefined appropriately.

This plugin mostly works, but there are two major remaining issues.
Therefore, this plugin is disabled by default and has zero priority.

(1) Mouse is not implemented. The stupid fix involves cut&paste from
the XCB-GLX plugin. Moving mouse event reception to the window provider
might be smart though: I would rather not corrupt the EGL plugin with
X11, otherwise we might as well stick to GLX anyway.

(2) Resize events are mostly not working. EGL does not provide
notification for this. But there may be some bugs in Control() too.