Merge VoD module into RTP stream output module
Cleaner and better working solution for VoD than the bad hack that we
currently have. Instead of (badly) duplicating large portions of codes,
it reuses the existing RTSP stack and factorizes the RTP payload code,
and provides support for a number of ES not limited to one audio and one
video tracks, as well as for missing features in RTSP headers.

The VoD module is responsible for starting and running the RTSP stack.
When a VoD media instance is started, the VLM passes down media and
session parameters as variables to the RTP stream output object, that
uses them to hook back into the VoD code (and through it, into the RTSP
stack), to retrieve accurate, already set up parameters, and provide
playback data.